requirements.txt
Django==3.1.3
django-filter==21.1
djangorestframework==3.13.1
djangorestframework-jsonapi==5.0.0